arm: fix TB alignment check
Sometimes page faults happen during the translation of the target instructions.
To avoid the faults in the middle of the TB we have to stop translation at
the end of the page. Current implementation of ARM translation assumes that
instructions are aligned to their own size (4 or 2 bytes). But in thumb2 mode
4-byte instruction can be aligned to 2 bytes. In some cases such an alignment
leads to page fault.
This patch adds check that allows translation of such instructions only in
the beginning of the TB.